fix(rest-api): complete review hardening
